refactor: rename rule file
