+++ src/__tests__/linter.test.ts
@@ -4,9 +4,9 @@ import * as fs from 'fs';
 import { cloneDeep } from 'lodash';
 import { mergeRulesets } from '../rulesets/merger';
 import { oas2Functions } from '../rulesets/oas2';
-import * as oas2Ruleset from '../rulesets/oas2/index.json';
+import * as oas2Ruleset from '../rulesets/oas2/rules.json';
 import { oas3Functions } from '../rulesets/oas3';
-import * as oas3Ruleset from '../rulesets/oas3/index.json';
+import * as oas3Ruleset from '../rulesets/oas3/rules.json';
 import { Spectral } from '../spectral';
 import { RuleCollection } from '../types';
 import { IRulesetFile } from '../types/ruleset';
