+++ src/rulesets/oas/functions/oasOpSecurityDefined/__tests__/index.ts
@@ -1,8 +1,8 @@
 import { RuleType, Spectral } from '../../../../../index';
 import { oas2Functions } from '../../../../oas2/index';
-import { rules as oas2Rules } from '../../../../oas2/index.json';
+import { rules as oas2Rules } from '../../../../oas2/rules.json';
 import { oas3Functions } from '../../../../oas3/index';
-import { rules as oas3Rules } from '../../../../oas3/index.json';
+import { rules as oas3Rules } from '../../../../oas3/rules.json';
 
 const oas2Ruleset = { functions: oas2Functions(), rules: oas2Rules };
 const oas3Ruleset = { functions: oas3Functions(), rules: oas3Rules };
